anki-writer/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
httpx>=0.24.0
python-dotenv>=1.0.0
anki==25.9.2
